+++ src/main/resources/ci3-stubs/language_helper.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * CodeIgniter 3 Language Helper (stub for IDE).
+ * Load with: $this->load->helper('language');
+ * @see https://codeigniter.com/userguide3/helpers/language_helper.html
+ */
+
+/**
+ * Returns a line of text from a loaded language file. Optionally wraps it in an HTML label.
+ *
+ * @param string $line       Language line key
+ * @param string $for        HTML "for" attribute (ID of the element we're creating a label for). If not empty, output is wrapped in <label for="...">.
+ * @param array  $attributes Any additional HTML attributes for the label (default: array())
+ * @return string The language line; or in an HTML label tag if $for is not empty
+ */
+function lang($line, $for = '', $attributes = array()) {}